Move env modifying tests to a separate integration test
